refactor permissions out of the credentials service
- permissions are now in credhub backends at the handler level
- there is no longer a dependency from credentials to permissions

[#165568088]

+++ backends/credhub/src/main/kotlin/org/cloudfoundry/credhub/certificates/DefaultCertificatesHandler.kt
@@ -2,40 +2,53 @@ package org.cloudfoundry.credhub.certificates
 
 import com.google.common.collect.Lists
 import org.cloudfoundry.credhub.ErrorMessages
+import org.cloudfoundry.credhub.PermissionOperation
+import org.cloudfoundry.credhub.PermissionOperation.DELETE
+import org.cloudfoundry.credhub.PermissionOperation.READ
+import org.cloudfoundry.credhub.PermissionOperation.WRITE
 import org.cloudfoundry.credhub.audit.AuditableCredential
 import org.cloudfoundry.credhub.audit.CEFAuditRecord
+import org.cloudfoundry.credhub.auth.UserContextHolder
 import org.cloudfoundry.credhub.credential.CertificateCredentialValue
 import org.cloudfoundry.credhub.domain.CertificateCredentialVersion
 import org.cloudfoundry.credhub.domain.CredentialVersion
 import org.cloudfoundry.credhub.entity.Credential
 import org.cloudfoundry.credhub.exceptions.EntryNotFoundException
+import org.cloudfoundry.credhub.exceptions.PermissionException
 import org.cloudfoundry.credhub.generate.GenerationRequestGenerator
 import org.cloudfoundry.credhub.generate.UniversalCredentialGenerator
-import org.cloudfoundry.credhub.permissions.PermissionedCertificateService
 import org.cloudfoundry.credhub.requests.CertificateRegenerateRequest
 import org.cloudfoundry.credhub.requests.CreateVersionRequest
 import org.cloudfoundry.credhub.requests.UpdateTransitionalVersionRequest
+import org.cloudfoundry.credhub.services.DefaultCertificateService
+import org.cloudfoundry.credhub.services.PermissionCheckingService
 import org.cloudfoundry.credhub.views.CertificateCredentialView
 import org.cloudfoundry.credhub.views.CertificateCredentialsView
 import org.cloudfoundry.credhub.views.CertificateVersionView
 import org.cloudfoundry.credhub.views.CertificateView
 import org.cloudfoundry.credhub.views.CredentialView
+import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Service
+import java.util.ArrayList
 import java.util.UUID
 
 @Service
 class DefaultCertificatesHandler(
-    private val permissionedCertificateService: PermissionedCertificateService,
-    private val certificateService: CertificateService,
+    private val certificateService: DefaultCertificateService,
     private val credentialGenerator: UniversalCredentialGenerator,
     private val generationRequestGenerator: GenerationRequestGenerator,
-    private val auditRecord: CEFAuditRecord
+    private val auditRecord: CEFAuditRecord,
+    private val permissionCheckingService: PermissionCheckingService,
+    private val userContextHolder: UserContextHolder,
+    @Value("\${security.authorization.acls.enabled}") private val enforcePermissions: Boolean
 ) : CertificatesHandler {
 
     override fun handleRegenerate(
         credentialUuid: String,
         request: CertificateRegenerateRequest
     ): CredentialView {
+
+        checkPermissionsByUuid(credentialUuid, WRITE)
 
         val existingCredentialVersion = certificateService
             .findByCredentialUuid(credentialUuid)
@@ -44,9 +57,10 @@ class DefaultCertificatesHandler(
             .createGenerateRequest(existingCredentialVersion)
         val credentialValue = credentialGenerator
             .generate(generateRequest) as CertificateCredentialValue
+
         credentialValue.isTransitional = request.isTransitional
 
-        val credentialVersion = permissionedCertificateService
+        val credentialVersion = certificateService
             .save(
                 existingCredentialVersion,
                 credentialValue,
@@ -59,33 +73,42 @@ class DefaultCertificatesHandler(
     }
 
     override fun handleGetAllRequest(): CertificateCredentialsView {
-        val credentialList = permissionedCertificateService.getAll()
+        val credentialList = filterPermissions(certificateService.getAll())
         val list = convertCertificateCredentialsToCertificateCredentialViews(credentialList)
+
         auditRecord.addAllCredentials(Lists.newArrayList<AuditableCredential>(credentialList))
+
         return CertificateCredentialsView(list)
     }
 
     override fun handleGetByNameRequest(name: String): CertificateCredentialsView {
-        val credentialList = permissionedCertificateService.getByName(name)
+        checkPermissionsByName(name, READ)
+
+        val credentialList = certificateService.getByName(name)
         val list = convertCertificateCredentialsToCertificateCredentialViews(credentialList)
+
         return CertificateCredentialsView(list)
     }
 
-    override fun handleGetAllVersionsRequest(uuidString: String, current: Boolean): List<CertificateView> {
+    override fun handleGetAllVersionsRequest(certificateId: String, current: Boolean): List<CertificateView> {
+
         val uuid: UUID
         try {
-            uuid = UUID.fromString(uuidString)
+            uuid = UUID.fromString(certificateId)
         } catch (e: IllegalArgumentException) {
             throw EntryNotFoundException(ErrorMessages.Credential.INVALID_ACCESS)
         }
 
-        val credentialList = permissionedCertificateService.getVersions(uuid, current)
+        checkPermissionsByUuid(uuid.toString(), READ)
+        val credentialList = certificateService.getVersions(uuid, current)
 
         return credentialList.map { credential -> CertificateView(credential as CertificateCredentialVersion) }
     }
 
     override fun handleDeleteVersionRequest(certificateId: String, versionId: String): CertificateView {
-        val deletedVersion = permissionedCertificateService
+        checkPermissionsByUuid(certificateId, DELETE)
+
+        val deletedVersion = certificateService
             .deleteVersion(UUID.fromString(certificateId), UUID.fromString(versionId))
         return CertificateView(deletedVersion)
     }
@@ -94,6 +117,7 @@ class DefaultCertificatesHandler(
         certificateId: String,
         requestBody: UpdateTransitionalVersionRequest
     ): List<CertificateView> {
+        checkPermissionsByUuid(certificateId, WRITE)
         var versionUUID: UUID? = null
 
         if (requestBody.versionUuid != null) {
@@ -101,7 +125,7 @@ class DefaultCertificatesHandler(
         }
 
         val credentialList: List<CredentialVersion>
-        credentialList = permissionedCertificateService
+        credentialList = certificateService
             .updateTransitionalVersion(UUID.fromString(certificateId), versionUUID)
 
         return credentialList
@@ -109,9 +133,11 @@ class DefaultCertificatesHandler(
     }
 
     override fun handleCreateVersionsRequest(certificateId: String, requestBody: CreateVersionRequest): CertificateView {
+        checkPermissionsByUuid(certificateId, WRITE)
+
         val certificateCredentialValue = requestBody.value
         certificateCredentialValue.isTransitional = requestBody.isTransitional
-        val credentialVersion = permissionedCertificateService.set(
+        val credentialVersion = certificateService.set(
             UUID.fromString(certificateId),
             certificateCredentialValue
         )
@@ -121,7 +147,7 @@ class DefaultCertificatesHandler(
 
     private fun convertCertificateCredentialsToCertificateCredentialViews(certificateCredentialList: List<Credential>): List<CertificateCredentialView> {
         return certificateCredentialList.map { credential ->
-            val certificateVersions = permissionedCertificateService.getAllValidVersions(credential.uuid!!) as List<CertificateCredentialVersion>
+            val certificateVersions = certificateService.getAllValidVersions(credential.uuid!!) as List<CertificateCredentialVersion>
 
             var signedBy = ""
             if (certificateVersions.isNotEmpty()) {
@@ -133,7 +159,7 @@ class DefaultCertificatesHandler(
             }
 
             val signedCertificates = if (credential.name != null) {
-                permissionedCertificateService.findSignedCertificates(credential.name!!)
+                certificateService.findSignedCertificates(credential.name!!)
             } else {
                 emptyList()
             }
@@ -148,5 +174,75 @@ class DefaultCertificatesHandler(
 
             CertificateCredentialView(credential.name, credential.uuid, certificateVersionViews, signedBy, signedCertificates)
         }
+    }
+
+    private fun checkPermissionsByName(name: String, permissionOperation: PermissionOperation) {
+        if (!enforcePermissions) return
+
+        if (!permissionCheckingService.hasPermission(
+                userContextHolder.userContext.actor!!,
+                name,
+                permissionOperation
+            )) {
+            if (permissionOperation == WRITE) {
+                throw PermissionException(ErrorMessages.Credential.INVALID_ACCESS)
+            } else {
+                throw EntryNotFoundException(ErrorMessages.Credential.INVALID_ACCESS)
+            }
+        }
+    }
+
+    private fun checkPermissionsByUuid(uuid: String, permissionOperation: PermissionOperation) {
+        if (!enforcePermissions) return
+
+        val certificate = certificateService.findByCredentialUuid(uuid)
+
+        if (!permissionCheckingService.hasPermission(
+                userContextHolder.userContext.actor!!,
+                certificate.name,
+                permissionOperation
+            )) {
+            if (permissionOperation == WRITE) {
+                throw PermissionException(ErrorMessages.Credential.INVALID_ACCESS)
+            } else {
+                throw EntryNotFoundException(ErrorMessages.Credential.INVALID_ACCESS)
+            }
+        }
+    }
+
+    private fun filterPermissions(unfilteredCredentials: List<Credential>): List<Credential> {
+        if (!enforcePermissions) {
+            return unfilteredCredentials
+        }
+        val actor = userContextHolder.userContext.actor
+        val paths = permissionCheckingService.findAllPathsByActor(actor)
+
+        if (paths.contains("/*")) return unfilteredCredentials
+        if (paths.isEmpty()) return ArrayList()
+
+        val filteredCredentials = ArrayList<Credential>()
+
+        for (credential in unfilteredCredentials) {
+            val credentialName = credential.name
+            if (paths.contains(credentialName)) {
+                filteredCredentials.add(credential)
+            }
+
+            val result = ArrayList<String>()
+
+            for (i in 1 until credentialName!!.length) {
+                if (credentialName[i] == '/') {
+                    result.add(credentialName.substring(0, i) + "/*")
+                }
+            }
+
+            for (credentialPath in result) {
+                if (paths.contains(credentialPath)) {
+                    filteredCredentials.add(credential)
+                    break
+                }
+            }
+        }
+        return filteredCredentials
     }
 }
